+++ ent/lang/lang_test.go
@@ -9,11 +9,12 @@ import (
 
 func TestString(t *testing.T) {
 	assert.Equal(t, lang.English.String(), "eng")
-	assert.Equal(t, lang.Default.String(), "eng")
+	assert.Equal(t, lang.German.String(), "deu")
+	assert.Equal(t, lang.None.String(), "")
 }
 
 func TestLanguageSet(t *testing.T) {
-	ls := lang.LanguagesSet()
+	ls := lang.LanguagesSet
 	_, ok := ls[lang.English]
 	assert.True(t, ok)
 	_, ok = ls[lang.German]
@@ -42,6 +43,6 @@ func TestDetectLang(t *testing.T) {
 	// unknown language detection
 	text = "Однажды в студеную, зимнюю пору я из лесу вышел"
 	l, code = lang.DetectLanguage([]rune(text))
-	assert.Equal(t, l, lang.Default)
+	assert.Equal(t, l, lang.English)
 	assert.Equal(t, code, "rus")
 }