+++ ent/lang/language.go
@@ -2,6 +2,7 @@ package lang
 
 import (
 	"fmt"
+	"sort"
 
 	"github.com/abadojack/whatlanggo"
 )
@@ -12,49 +13,68 @@ type Language int
 // Set of languages. Last one is an indicator of the 'edge', as well as
 // a default value for GnFinder.Language.
 const (
-	Default Language = iota
+	None Language = iota
 	English
 	German
-	NotSet
 )
 
-func (l Language) String() string {
-	languages := [...]string{"eng", "eng", "deu", ""}
-	return languages[l]
+var langMap = map[Language]string{
+	None:    "",
+	English: "eng",
+	German:  "deu",
 }
 
-// New takes a string and returns a matching language. If language could not
-// be found by the string, the function returns lang.DefaultLanguage and an
-// error.
-func New(lang string) (Language, error) {
-	for _, l := range SupportedLanguages() {
-		if l.String() == lang {
-			return l, nil
+var langStrMap = func() map[string]Language {
+	res := make(map[string]Language)
+	for k, v := range langMap {
+		if v != "" {
+			res[v] = k
 		}
 	}
-	var l Language
-	return l, fmt.Errorf("unknown language %s", lang)
-}
+	return res
+}()
 
-// SupportedLanguages returns a slice of supported by gnfinder languages.
-func SupportedLanguages() []Language {
-	var res []Language
-	for i := 1; i < int(NotSet); i++ {
-		l := Language(i)
-		res = append(res, l)
+// LanguagesSet contains supported languages and their string representation.
+var LanguagesSet = func() map[Language]string {
+	res := make(map[Language]string)
+	for k, v := range langMap {
+		if k != None {
+			res[k] = v
+		}
 	}
 	return res
+}()
+
+func (l Language) String() string {
+	return langMap[l]
 }
 
-// LanguagesSet returns a 'set' of languages for more effective
-// lookup of a language.
-func LanguagesSet() map[Language]struct{} {
-	var empty struct{}
-	ls := make(map[Language]struct{})
-	for i := 0; i < int(NotSet); i++ {
-		ls[Language(i)] = empty
+// New takes a string and returns a matching language. If string is "detect",
+// returns lang.None. If language could not be found by the string, the
+// function returns lang.English and an error.
+func New(s string) (Language, error) {
+	switch s {
+	case "":
+		return English, nil
+	case "detect":
+		return None, nil
 	}
-	return ls
+
+	if l, ok := langStrMap[s]; ok {
+		return l, nil
+	}
+	return English, fmt.Errorf("unknown language %s", s)
+}
+
+func LangStrings() []string {
+	res := make([]string, 0, 10)
+	for k := range langStrMap {
+		if k != "" {
+			res = append(res, k)
+		}
+	}
+	sort.Strings(res)
+	return res
 }
 
 // DetectLanguage finds the most probable language for a text.
@@ -71,6 +91,6 @@ func DetectLanguage(text []rune) (Language, string) {
 	case "deu":
 		return German, code
 	default:
-		return Default, code
+		return English, code
 	}
 }